+++ src/Components/areas/DemandArea.tsx
@@ -1,8 +1,8 @@
 import React, { useContext } from "react";
-import { HouseDemandContext } from "./HouseDemandContext";
-import Title from "./Title";
-import IncrementAndToggle from "./IncrementAndToggle";
-import FoodChainMagnatePalette from "./ColorPalette";
+import { HouseDemandContext } from "../HouseDemandContext";
+import Title from "../atoms/Title";
+import IncrementAndToggle from "../atoms/IncrementAndToggle";
+import FoodChainMagnatePalette from "../ColorPalette";
 
 const DemandArea: React.FC = () => {
   const { demand, setDemand, milestones, setMilestones } = useContext(HouseDemandContext);
