+++ src/Components/areas/index.ts
@@ -1,0 +1,3 @@
+export { default as UnitPriceArea } from "./UnitPriceArea";
+export { default as DemandArea } from "./DemandArea";
+export { default as RevenueBonusArea } from "./RevenueBonusArea";
